// src/actions/domain/enums.ts
/**
 * Auto generated. DO NOT edit manually.
 * Last updated on: Thu, 27 Jul 2023 15:06:33 GMT
 */

/**
 * @enum
 * @readonly
 * @description PropWorlds Action Names
 */
export enum PropWorldsActionName {
    Arbapprove = 'arbapprove',
    Arbdeny = 'arbdeny',
    Cancelprop = 'cancelprop',
    Cancelwip = 'cancelwip',
    Clearexpprop = 'clearexpprop',
    Comment = 'comment',
    Completework = 'completework',
    Createprop = 'createprop',
    Delegatecat = 'delegatecat',
    Delegatevote = 'delegatevote',
    Dispute = 'dispute',
    Finalize = 'finalize',
    Startwork = 'startwork',
    Undelegateca = 'undelegateca',
    Updateconfig = 'updateconfig',
    Updpropvotes = 'updpropvotes',
    Voteprop = 'voteprop',
    Votepropfin = 'votepropfin',
};